commit 46  - part 2

+++ src/pt/ulusofona/lp2/deisichess/Statistics.kt
@@ -31,7 +31,7 @@ fun top3MaisBaralhadas(gameManager: GameManager): List<String> {
     val pieces = gameManager.getGameStatus().getTheBoard().getAllPieces()
     val filtered = pieces.values.filter { it.getInvalidMoves() > 0 }
     val sortedPieces = filtered.sortedWith(
-        compareByDescending<Piece> { it.getInvalidMoves() / (it.getInvalidMoves() + it.getValidMoves()) } .thenBy { it.getName() }
+        compareByDescending<Piece> { (it.getInvalidMoves()).toDouble() / (it.getInvalidMoves() + it.getValidMoves()).toDouble() } .thenBy { it.getName() }
     )
     val listResult = sortedPieces.map { "${it.getTeam()}:${it.getName()}:${it.getInvalidMoves()}:${it.getValidMoves()}" }
     return listResult.take(3)
